Rename : 폴더 구조 변경
- 폴더 구조 가독성을 위해 Route Group, Private Folder 적용
