+++ src/app/(route)/register/_components/RegisterForm.module.scss
@@ -1,4 +1,4 @@
-@use '@/app/styles/themes.scss' as themes;
+@use '@/app/_styles/themes.scss' as themes;
 
 $textAreaHieght: calc(themes.$fontSize * 10);
 
@@ -192,8 +192,8 @@ $textAreaHieght: calc(themes.$fontSize * 10);
         background-color: themes.$subColor;
 
         .profileImage {
-          width : 100%;
-          height : 100%;
+          width: 100%;
+          height: 100%;
           object-fit: cover;
         }
       }
