+++ src/app/(route)/register/page.module.scss
@@ -1,4 +1,4 @@
-@use '@/app/styles/themes.scss' as themes;
+@use '@/app/_styles/themes.scss' as themes;
 
 .container {
   max-width: themes.$smallContentMaxWidth;
