+++ src/app/_components/BannerList.module.scss
@@ -1,4 +1,4 @@
-@use '@/app/styles/themes.scss' as themes;
+@use '@/app/_styles/themes.scss' as themes;
 
 // 비행기 개수, 비율
 $airplaneAmount: 20;
