+++ src/app/_components/RecommendedList.module.scss
@@ -1,4 +1,4 @@
-@use '@/app/styles/themes.scss' as themes;
+@use '@/app/_styles/themes.scss' as themes;
 
 .recommendedContainer {
   display: flex;
