+++ src/app/layout.module.scss
@@ -1,4 +1,4 @@
-@use '@/app/styles/themes.scss' as themes;
+@use '@/app/_styles/themes.scss' as themes;
 
 // CSS 컨벤션을 따를 것
 .header {
